feat: show multiple ingestions ui only if gateway is enabled
