+++ frontend/src/constants/features.ts
@@ -20,4 +20,5 @@ export enum FeatureKeys {
 	ONBOARDING = 'ONBOARDING',
 	CHAT_SUPPORT = 'CHAT_SUPPORT',
 	PLANNED_MAINTENANCE = 'PLANNED_MAINTENANCE',
+	GATEWAY = 'GATEWAY',
 }
